+++ tests/e2e/cases/fs_promises_stat_access_ignored_arguments/in.ts
@@ -15,18 +15,30 @@ fs.promises.access(file, void 0, mark("access-void-default"));
 fs.promises.access(file, defaultMode, mark("access-alias-default"));
 fs.promises.access(file, fs.constants.F_OK, mark("access-mode"));
 
+let statIsFile = false;
+let statSize = 0;
+let lstatIsFile = false;
+let lstatSize = 0;
+let missing = false;
+
 fs.promises.stat(file, void 0, mark("stat")).then((value: FSStats): void => {
-    console.log("stat:", value.isFile(), value.size);
+    statIsFile = value.isFile();
+    statSize = value.size;
 });
 
 fs.promises.lstat(file, void 0, mark("lstat")).then((value: FSStats): void => {
-    console.log("lstat:", value.isFile(), value.size);
+    lstatIsFile = value.isFile();
+    lstatSize = value.size;
 });
 
 fs.promises.stat(file + ".missing", { throwIfNoEntry: false }, mark("missing")).then((value: FSStats | undefined): void => {
-    console.log("missing:", value === undefined);
+    missing = value === undefined;
 });
 
-fs.rmSync(file, { force: true });
-
 console.log("events:", events.join("|"));
+setImmediate((): void => {
+    console.log("stat:", statIsFile, statSize);
+    console.log("lstat:", lstatIsFile, lstatSize);
+    console.log("missing:", missing);
+    fs.rmSync(file, { force: true });
+});